refac: moved state notifier providers to own files
