+++ lib/providers/state_notifiers/theme_state_notifier.dart
@@ -1,0 +1,14 @@
+import 'package:apidash/services/hive_services.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class ThemeStateNotifier extends StateNotifier<bool> {
+  ThemeStateNotifier(this.hiveHandler) : super(false) {
+    state = hiveHandler.getDarkMode() ?? false;
+  }
+  final HiveHandler hiveHandler;
+
+  Future<void> toggle() async {
+    state = !state;
+    await hiveHandler.setDarkMode(state);
+  }
+}
